CAN-TP implementation started

+++ CanInterface.cpp
@@ -1,41 +1,51 @@
-#include"CanInterface.h"
+#include "CanInterface.h"
 
-void CanInterface::SetSendFunction(SendFunctionType sendFunction)
+void
+CanInterface::SetSendFunction (SendFunctionType sendFunction)
 {
-    _sendFunction = sendFunction;
+  _sendFunction = sendFunction;
 }
 
-void CanInterface::SetReceiveFunction(ReceiveFunctionType receiveFunction)
+void
+CanInterface::SetReceiveFunction (ReceiveFunctionType receiveFunction)
 {
-    _receiveFunction = receiveFunction;
+  _receiveFunction = receiveFunction;
 }
 
-void CanInterface::SetTimeoutFunction(SetTimeoutFunctionType setTimeoutFunction)
+void
+CanInterface::SetTimeoutFunction (SetTimeoutFunctionType setTimeoutFunction)
 {
-    _setTimeoutFunction = setTimeoutFunction;
+  _setTimeoutFunction = setTimeoutFunction;
 }
 
-uint32_t CanInterface::SendMessage(uint8_t* buffer, uint32_t length, uint32_t destination)
+uint32_t
+CanInterface::SendMessage (uint8_t *buffer, uint8_t length,
+                           uint32_t destination)
 {
-    return _sendFunction(buffer, length, destination);
+  return _sendFunction (buffer, length, destination);
 }
 
-uint32_t CanInterface::ReceiveMessage(uint8_t* buffer, uint32_t* length, uint32_t source)
+uint32_t
+CanInterface::ReceiveMessage (uint8_t *buffer, uint8_t *length,
+                              uint32_t source)
 {
-    return _receiveFunction(buffer, length, source);
+  return _receiveFunction (buffer, length, source);
 }
 
-void CanInterface::SetTimeout(uint32_t timeout)
+void
+CanInterface::SetTimeout (uint32_t timeout)
 {
-    _setTimeoutFunction(timeout);
+  _setTimeoutFunction (timeout);
 }
 
-bool CanInterface::IsCanFd()
+bool
+CanInterface::IsCanFd ()
 {
-    return _canFd;
+  return _canFd;
 }
 
-void CanInterface::SetCanFd(bool canFd)
+void
+CanInterface::SetCanFd (bool canFd)
 {
-    _canFd = canFd;
+  _canFd = canFd;
 }
